+++ src/MainLogic/Main.py
@@ -10,7 +10,7 @@ import os
 import threading
 import rclpy
 from rclpy.executors import MultiThreadedExecutor
-from MainLogic.Lib import rosBridgeNode as ros_bridge_module
+from MainLogic.core import ros_bridge_node as ros_bridge_module
 
 # 设置多进程启动方法为 'spawn'，确保子进程完全独立，不共享 ROS2 上下文
 multiprocessing.set_start_method('spawn', force=True)
@@ -54,7 +54,7 @@ def main():
     #创建ROS2节点与多线程执行器
     executor = MultiThreadedExecutor()
     '''需要用命名空间来保证修改修改的是全局变量'''
-    '''原来的from MainLogic.Lib.rosBridgeNode import rosBridgeNode,RosBridgeNodeInstance'''
+    '''原来的from MainLogic.core.ros_bridge_node import rosBridgeNode,RosBridgeNodeInstance'''
     '''是在本地命名空间里创建了RosBridgeNodeInstance,修改的是本地的RosBridgeNodeInstance,而不是全局的RosBridgeNodeInstance'''
     '''另外如果在本地命名空间创建全局变量要赋值的话,需要global关键字声明'''
     # ros_bridge_module.RosBridgeNodeInstance = ros_bridge_module.rosBridgeNode()
